refactor: 119 - move parameters to models package
