+++ lib/models/offre_emploi_filtres_parameters.dart
@@ -1,0 +1,21 @@
+class OffreEmploiSearchParametersFiltres {
+  final int? distance;
+
+  OffreEmploiSearchParametersFiltres._({
+    this.distance,
+  });
+
+  factory OffreEmploiSearchParametersFiltres.withFiltres({
+    int? distance,
+  }) {
+    return OffreEmploiSearchParametersFiltres._(
+      distance: distance,
+    );
+  }
+
+  factory OffreEmploiSearchParametersFiltres.noFiltres() {
+    return OffreEmploiSearchParametersFiltres._(
+      distance: null,
+    );
+  }
+}
