+++ lib/redux/actions/offre_emploi_actions.dart
@@ -1,6 +1,6 @@
 import 'package:pass_emploi_app/models/location.dart';
 import 'package:pass_emploi_app/models/offre_emploi.dart';
-import 'package:pass_emploi_app/redux/states/offre_emploi_search_parameters_state.dart';
+import 'package:pass_emploi_app/models/offre_emploi_filtres_parameters.dart';
 
 abstract class OffreEmploiAction {}
