+++ lib/redux/states/offre_emploi_search_parameters_state.dart
@@ -1,4 +1,5 @@
 import 'package:pass_emploi_app/models/location.dart';
+import 'package:pass_emploi_app/models/offre_emploi_filtres_parameters.dart';
 
 abstract class OffreEmploiSearchParametersState {
   OffreEmploiSearchParametersState._();
@@ -27,26 +28,4 @@ class OffreEmploiSearchParametersInitializedState extends OffreEmploiSearchParam
 
 class OffreEmploiSearchParametersStateNotInitializedState extends OffreEmploiSearchParametersState {
   OffreEmploiSearchParametersStateNotInitializedState() : super._();
-}
-
-class OffreEmploiSearchParametersFiltres {
-  final int? distance;
-
-  OffreEmploiSearchParametersFiltres._({
-    this.distance,
-  });
-
-  factory OffreEmploiSearchParametersFiltres.withFiltres({
-    int? distance,
-  }) {
-    return OffreEmploiSearchParametersFiltres._(
-      distance: distance,
-    );
-  }
-
-  factory OffreEmploiSearchParametersFiltres.noFiltres() {
-    return OffreEmploiSearchParametersFiltres._(
-      distance: null,
-    );
-  }
 }
